feat: translations deutsch (#285)
* Update de.json

I translated the stuff that wasn't translated. I changed everything from the formal "Sie" to the proper "du" because it was a) mixed and b) "Sie" is a formal thing to do that is weird to look at in gaming. Once the changes are approved, I can go over the site and look over it to see if I mistranslated some things. Otherwise you should be good to go for now.

* translations

few new changes, after looking at the page some more.