fix: index card drag and drop types (#408)
The Drag and Drop context needs an id to know where things can be dropped or not. In this case, when an index card was rendering the DnD context for sub cards it was using the id of the current card instead of the parent one.

I also added a check in the controllers to make sure to only move the cards if a card is found in the current array.