Fix CI requiring original repo (#560)
## ✅ Changes

<!-- Use prefixes: **chore**, **docs**, **feat**, **fix**, **refactor**,
**style** or **test** -->

- **fix**: This should fix the CI failing on forks.

## 🌄 Context

I had noticed when fixing the issue of the cap on Numeric fields that
the CI consistently failed. Looking from the build logs, I think the
issue is that it looks for secrets related to the main repo in a way
that seems hardcoded. This _should_ fix that particular issue.